permission seed and authorization added

+++ app/routes/product_routes.js
@@ -4,41 +4,47 @@ import fs from "fs";
 import path from "path";
 import upload from "../middlewares/image_middleware.js";
 import { validateProductReqBody } from "../validators/validate_req_body.js";
-import ErrorConstants from "../constant/errors.js";
+import { Permission } from "../constant/authorization.js";
+import authorize from "../middlewares/middleware.js";
 
 const router = Router();
 
-router.get("/product", async (req, res) => {
-  const results = await prisma.product.findMany({
-    include: {
-      Category: {
-        select: {
-          name: true,
+router.get(
+  "/product",
+  authorize(Permission.BROWSE_PRODUCTS),
+  async (req, res) => {
+    const results = await prisma.product.findMany({
+      include: {
+        Category: {
+          select: {
+            name: true,
+          },
         },
-      },
-      ProductImage: {
-        select: {
-          image_url: true,
+        ProductImage: {
+          select: {
+            image_url: true,
+          },
         },
-      },
-      ProductSize: {
-        select: {
-          Size: {
-            select: {
-              name: true,
+        ProductSize: {
+          select: {
+            Size: {
+              select: {
+                name: true,
+              },
             },
           },
         },
       },
-    },
-  });
-  res.json(results);
-});
+    });
+    res.json(results);
+  }
+);
 
 router.post(
   "/product",
   upload.single("image"),
   validateProductReqBody,
+  authorize(Permission.ADD_PRODUCTS),
   async (req, res) => {
     const { name, description, price, category_id, quantity, rating } =
       req.body;
@@ -102,6 +108,7 @@ router.put(
   "/product/:id",
   upload.single("image"),
   validateProductReqBody,
+  authorize(Permission.EDIT_PRODUCTS),
   async (req, res) => {
     const rootUrl = `${req.protocol}://${req.get("host")}`;
     if (isNaN(req.params.id)) {
@@ -138,24 +145,28 @@ router.put(
   }
 );
 
-router.delete("/product/:id", async (req, res) => {
-  if (isNaN(req.params.id)) {
-    res.status(400).json({ message: "Invalid ID" });
-  } else {
-    const product_id = await prisma.product.findFirst({
-      where: { id: Number(req.params.id) },
-    });
-    if (!product_id) {
-      res.status(404).json({ message: "Product Not Found" });
+router.delete(
+  "/product/:id",
+  authorize(Permission.DELETE_PRODUCTS),
+  async (req, res) => {
+    if (isNaN(req.params.id)) {
+      res.status(400).json({ message: "Invalid ID" });
     } else {
-      const product_id = await prisma.productImage.findFirst({
-        where: { product_id: Number(req.params.id) },
+      const product_id = await prisma.product.findFirst({
+        where: { id: Number(req.params.id) },
       });
-      await prisma.productImage.delete({ where: { id: product_id.id } });
-      await prisma.product.delete({ where: { id: Number(req.params.id) } });
-      res.status(200).json({ message: "Product has been deleted" });
+      if (!product_id) {
+        res.status(404).json({ message: "Product Not Found" });
+      } else {
+        const product_id = await prisma.productImage.findFirst({
+          where: { product_id: Number(req.params.id) },
+        });
+        await prisma.productImage.delete({ where: { id: product_id.id } });
+        await prisma.product.delete({ where: { id: Number(req.params.id) } });
+        res.status(200).json({ message: "Product has been deleted" });
+      }
     }
   }
-});
+);
 
 export default router;
